Refactoring, introduced package smile for fouriermethod.products.
Note: this will be subject of further refactorings.
