+++ src/main/java/net/finmath/fouriermethod/products/smile/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Products which are provide a "smile function" \( K \mapsto V(K) \) mapping a product strike to
+ * the corresponding product value.
+ * 
+ * @author Alessandro Gnoatto
+ * @author Christian Fries
+ */
+package net.finmath.fouriermethod.products.smile;
